set deployment params

+++ chrome_extension/background.js
@@ -1,4 +1,4 @@
-var SERVER = "http://localhost:5050";
+var SERVER = "http://automaticbrowser.com";
 var active = false;
 var current_url = "NONE";
 var user_id = null;
@@ -6,44 +6,46 @@ var key = null;
 var timeout = null;
 var auto = false;
 
+var verbose = false;
+if (verbose) {
+    console.log("Automatic Browser");
+    chrome.storage.sync.clear();
+}
+
 chrome.browserAction.setIcon({path: "icon_19_disable.png"});
 chrome.browserAction.setBadgeBackgroundColor({color: [255, 0, 0, 0]});
 
-// take this out, it's for testing only:
-console.log("hello world");
-chrome.storage.sync.clear();
-
 function initUser () {
-    console.log("initUser");
+    if (verbose) console.log("initUser");
     chrome.storage.sync.get("user_id", function(data) {
         if (data['user_id'] != null && data['key'] != null) {
             user_id = data['user_id'];
             key = data['key'];
-            console.log("user_id is " + user_id);
-            console.log("key is " + key);
+            if (verbose) console.log("user_id is " + user_id);
+            if (verbose) console.log("key is " + key);
         } else {
-            console.log("getting new user id");
+            if (verbose) console.log("getting new user id");
             $.post(SERVER, {'action': "new_user"}, function (data) {
                 user_id = data;
                 key = Math.random().toString(36).slice(2); // random 16-digit string
                 chrome.storage.sync.set({"user_id": data});
                 chrome.storage.sync.set({"key": key});
-                console.log("user_id is " + user_id);
-                console.log("key is " + key);
+                if (verbose) console.log("user_id is " + user_id);
+                if (verbose) console.log("key is " + key);
             });
         }
     });
 }
 
 function turnOn () {
-    console.log("background.turnOn");  
+    if (verbose) console.log("background.turnOn");  
     active = true;
     initUser();
     chrome.browserAction.setIcon({path: "icon_19.png"});
 }
 
 function turnOff () {
-    console.log("background.turnOff");  
+    if (verbose) console.log("background.turnOff");  
     active = false;
     if (timeout != null) clearTimeout(timeout);    
     chrome.browserAction.setIcon({path: "icon_19_disable.png"});
@@ -57,7 +59,7 @@ function checkUrl () {
     // console.log("background.checkUrl");  
     chrome.windows.getCurrent(function (window) {
         if (window == null || !window.focused) {
-            console.log('(using other application)');
+            if (verbose) console.log('(using other application)');
             current_url = 'NONE';
             postUrl();            
         } else {
@@ -65,7 +67,7 @@ function checkUrl () {
                 if (tab != null) {
                     if (tab.url != current_url) {
                         if (tab.url.substr(0, 4) != "http") {
-                            console.log("(settings page, ignoring)");                            
+                            if (verbose) console.log("(settings page, ignoring)");                            
                         } else {
                             current_url = tab.url;
                             postUrl();
@@ -75,7 +77,7 @@ function checkUrl () {
                         if (auto == true) auto = false;    
                     }
                 } else {
-                    console.log("(no tabs, ignoring)");
+                    if (verbose) console.log("(no tabs, ignoring)");
                 }
             });                        
         }
@@ -83,7 +85,7 @@ function checkUrl () {
 }
 
 function postUrl () {
-    console.log("background.postUrl " + current_url);
+    if (verbose) console.log("background.postUrl " + current_url);
     if (timeout != null) clearTimeout(timeout);    
     if (current_url != 'NONE') {
         var parts = parseURL(current_url);
@@ -97,7 +99,7 @@ function postUrl () {
     }
     $.post(SERVER, {action: 'report', user_id: user_id, host: host, page: page, auto: auto}, function(data) {
         if (data == "NOFUTURE") {
-            console.log("--> received NO FUTURE");
+            if (verbose) console.log("--> received NO FUTURE");
             return;
         }
         var parts = data.split(" ");
@@ -105,7 +107,7 @@ function postUrl () {
         var page = parts[1];
         var delay = parts[2];
         var url = "http://" + decrypt(host) + (page == '/' ? '/' : decrypt(page));
-        console.log("--> received future: " + url);
+        if (verbose) console.log("--> received future: " + url);
         timeout = setTimeout(function () {
             chrome.tabs.getSelected(null, function (tab) {
                 auto = true;
@@ -118,7 +120,7 @@ function postUrl () {
             });        
         }, delay);
     }).fail(function () {
-        console.log("post failed");
+        if (verbose) console.log("post failed");
     });
     if (auto == true) auto = false;
 }
